frontend/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.3
langchain-openai==0.1.7
langchain-pinecone==0.1.1
langsmith==0.1.17
tenacity==8.2.3
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
pinecone-client==3.2.2
flask==3.1.0
pydantic==1.10.13
transformers==4.46.3
torch==2.5.1
pinecone>=6.0.0